move arm commands into separate folder
