+++ src/main/java/frc/robot/RobotContainer.java
@@ -25,6 +25,12 @@ import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
 import edu.wpi.first.wpilibj2.command.StartEndCommand;
 import edu.wpi.first.wpilibj2.command.WaitCommand;
 import frc.robot.commands.*;
+import frc.robot.commands.Arm.ArmDown;
+import frc.robot.commands.Arm.ArmInCommand;
+import frc.robot.commands.Arm.ArmOutCommand;
+import frc.robot.commands.Arm.ArmUp;
+import frc.robot.commands.Arm.armDefualtControlCommand;
+import frc.robot.commands.Arm.armpidCommand;
 import frc.robot.commands.Drive.ArcadeDrive;
 import frc.robot.commands.Drive.SlowedWhileActiveCommand;
 import frc.robot.controllers.DriverController;
