+++ src/main/java/frc/robot/commands/Arm/ArmOutCommand.java
@@ -1,4 +1,4 @@
-package frc.robot.commands;
+package frc.robot.commands.Arm;
 
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.Subsystem;
@@ -6,18 +6,18 @@ import frc.robot.subsystems.ArmSubsystem;
 
 import java.util.Set;
 
-public class ArmInCommand implements Command {
+public class ArmOutCommand implements Command {
     private final ArmSubsystem armSubsystem;
     private final Set<Subsystem> subsystems;
 
-    public ArmInCommand(ArmSubsystem myArmSubsystem) {
+    public ArmOutCommand(ArmSubsystem myArmSubsystem) {
         this.armSubsystem = myArmSubsystem;
         this.subsystems = Set.of(this.armSubsystem);
     }
 
     @Override
     public void initialize() {
-        armSubsystem.telescopeArm(-0.95);
+        armSubsystem.telescopeArm(0.95);
     }
 
     @Override
