+++ src/main/java/frc/robot/commands/Arm/armpidCommand.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.Arm;
 
 import java.lang.ModuleLayer.Controller;
 import java.util.Set;
